requirements.txt
django==5.0
djangorestframework
django-filter
mock
Pillow
mysqlclient
python-dotenv
